Integration tests breakdown: Fix lint issues in the extracted integration_core tests
Move createAndroidHosts and createAndroidHostWithStorage into
integration_core_android_test.go, next to their only callers. Leaving them in
integration_core_test.go once the tests moved out made the unused linter report
them, failing CI. createAndroidHostForTest stays put since integration_mdm_test.go
also uses it.

Reattach doc comments displaced by the extraction. The move took each function
from its own declaration line onward, so a comment sitting above a function was
carried along with the preceding function instead. Reattach the 12 affected
comments to the functions they describe, including two left stranded at the end of
the previously extracted hosts and hosts_reports files.
