+++ server/service/integration_core_sessions_test.go
@@ -56,9 +56,6 @@ func (s *integrationTestSuite) TestSessionInfo() {
 	s.DoJSON("DELETE", fmt.Sprintf("/api/latest/fleet/sessions/%d", ssn.ID), nil, http.StatusNotFound, &delResp) // nolint:nilaway // createSession fails the test via require internally on error, cannot be nil here
 }
 
-// Free tier must not expose or accept fleet_desktop.sso_enabled, and must not
-// let a value stored under a premium license (before a downgrade) block
-// unrelated config changes.
 func (s *integrationTestSuite) TestLogLoginAttempts() {
 	t := s.T()
