+++ server/service/integration_core_test.go
@@ -388,6 +388,7 @@ const (
 }`
 )
 
+// Creates a set of results for use in tests for Query Results.
 func results(num int, hostID string) string {
 	b := strings.Builder{}
 	for i := range num {
@@ -411,39 +412,6 @@ func results(num int, hostID string) string {
 	}
 
 	return b.String()
-}
-
-func createAndroidHosts(t *testing.T, ds *mysql.Datastore, count int, teamID *uint) []uint {
-	ids := make([]uint, 0, count)
-	for i := range count {
-		host := &fleet.AndroidHost{
-			Host: &fleet.Host{
-				Hostname:       fmt.Sprintf("hostname%d", i),
-				ComputerName:   fmt.Sprintf("computer_name%d", i),
-				Platform:       "android",
-				OSVersion:      "Android 14",
-				Build:          fmt.Sprintf("build%d", i),
-				Memory:         1024,
-				TeamID:         teamID,
-				HardwareSerial: uuid.NewString(),
-			},
-			Device: &android.Device{
-				DeviceID:             strings.ReplaceAll(uuid.NewString(), "-", ""), // Remove dashes to fit in VARCHAR(37)
-				EnterpriseSpecificID: new(uuid.NewString()),
-				AppliedPolicyID:      new("1"),
-				LastPolicySyncTime:   new(time.Now().Add(-time.Hour)), // 1 hour ago
-			},
-		}
-		host.SetNodeKey(*host.Device.EnterpriseSpecificID)
-		ahost, err := ds.NewAndroidHost(context.Background(), host, false)
-		require.NoError(t, err)
-		ids = append(ids, ahost.Host.ID)
-	}
-	return ids
-}
-
-func createAndroidHostWithStorage(t *testing.T, ds *mysql.Datastore, teamID *uint) uint {
-	return createAndroidHostForTest(t, ds, teamID, false)
 }
 
 // createAndroidHostForTest creates an android host. companyOwned=false stores it as BYO
